refactor : User 타입 변경

+++ user-service/src/main/java/com/destiny/userservice/domain/entity/User.java
@@ -39,7 +39,7 @@ public class User extends BaseEntity {
     private String address2;
     private LocalDateTime birth;
 
-    private Double point;
+    private Long point;
     @Enumerated(EnumType.STRING)
     private MembershipGrade membershipGrade;
     private Long totalPrice;
